+++ packages/commerce-sdk-react/src/utils.test.ts
@@ -5,8 +5,9 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {resetDehydratedStateTimeStamp} from './utils'
 import {DehydratedState} from '@tanstack/react-query'
+import * as utils from './utils'
+
 describe('resetDehydratedStateTimeStamp', () => {
     const mockDehydratedState: DehydratedState = {
         mutations: [],
@@ -52,7 +53,7 @@ describe('resetDehydratedStateTimeStamp', () => {
 
     it('should update all dataUpdatedAt timestamps with the provided timestamp', () => {
         const specificTime = new Date(2023, 5, 15)
-        const result = resetDehydratedStateTimeStamp(mockDehydratedState, specificTime)
+        const result = utils.resetDehydratedStateTimeStamp(mockDehydratedState, specificTime)
 
         expect(result.queries[0].state.dataUpdatedAt).toBe(specificTime)
         expect(result.queries[1].state.dataUpdatedAt).toBe(specificTime)
@@ -61,7 +62,7 @@ describe('resetDehydratedStateTimeStamp', () => {
     it('should use current time when no timestamp is provided', () => {
         const before = Date.now()
 
-        const result = resetDehydratedStateTimeStamp(mockDehydratedState)
+        const result = utils.resetDehydratedStateTimeStamp(mockDehydratedState)
         const after = Date.now()
 
         expect(result.queries[0].state.dataUpdatedAt).toBeGreaterThanOrEqual(before)
@@ -74,7 +75,7 @@ describe('resetDehydratedStateTimeStamp', () => {
     it('should preserve all other properties of the state objects', () => {
         const specificTime = new Date(2023, 5, 15)
 
-        const result = resetDehydratedStateTimeStamp(mockDehydratedState, specificTime)
+        const result = utils.resetDehydratedStateTimeStamp(mockDehydratedState, specificTime)
 
         expect(result.queries[0].state.data).toBe('query data')
         expect(result.queries[0].queryKey).toEqual(['queries', 'key1'])
@@ -89,9 +90,37 @@ describe('resetDehydratedStateTimeStamp', () => {
             queries: []
         }
 
-        const result = resetDehydratedStateTimeStamp(emptyState, new Date(2023, 5, 15))
+        const result = utils.resetDehydratedStateTimeStamp(emptyState, new Date(2023, 5, 15))
 
         expect(result.mutations).toEqual([])
         expect(result.queries).toEqual([])
+    })
+})
+
+
+describe('Utils', () => {
+    test.each([
+        ['/callback', false],
+        ['https://pwa-kit.mobify-storefront.com/callback', true],
+        ['/social-login/callback', false]
+    ])('isAbsoluteUrl', (url, expected) => {
+        const isURL = utils.isAbsoluteUrl(url)
+        expect(isURL).toBe(expected)
+    })
+    test('extractCustomParameters only returns custom parameters', () => {
+        const parameters = {
+            c_param1: 'this is a custom',
+            param1: 'this is not a custom',
+            c_param2: 1,
+            param2: 2,
+            param3: false,
+            c_param3: true
+        }
+        const customParameters = utils.extractCustomParameters(parameters)
+        expect(customParameters).toEqual({
+            c_param1: 'this is a custom',
+            c_param2: 1,
+            c_param3: true
+        })
     })
 })