+++ packages/commerce-sdk-react/src/utils.ts
@@ -8,6 +8,7 @@
 import Cookies, {CookieAttributes} from 'js-cookie'
 import {IFRAME_HOST_ALLOW_LIST} from './constant'
 import {DehydratedState} from '@tanstack/react-query'
+import {helpers} from 'commerce-sdk-isomorphic'
 
 /** Utility to determine if you are on the browser (client) or not. */
 export const onClient = (): boolean => typeof window !== 'undefined'
@@ -138,4 +139,52 @@ export const resetDehydratedStateTimeStamp = (data: DehydratedState, timestamp?:
         mutations: data.mutations.map(updateQueryTimeStamp),
         queries: data.queries.map(updateQueryTimeStamp)
     }
+}
+/**
+ * Determines whether the given URL string is a valid absolute URL.
+ *
+ * Valid absolute URLs:
+ * - https://example.com
+ * - http://example.com
+ *
+ * Invalid or relative URLs:
+ * - http://example
+ * - example.com
+ * - /relative/path
+ *
+ * @param {string} url - The URL string to be checked.
+ * @returns {boolean} - Returns true if the given string is a valid absolute URL, false otherwise.
+ */
+export function isAbsoluteUrl(url: string): boolean {
+    return /^(https?:\/\/)/i.test(url)
+}
+
+/**
+ * Provides a platform-specific method for Base64 encoding.
+ *
+ * - In a browser environment (where `window` and `document` are defined),
+ *   the native `btoa` function is used.
+ * - In a non-browser environment (like Node.js), a fallback is provided
+ *   that uses `Buffer` to perform the Base64 encoding.
+ */
+export const stringToBase64 =
+    typeof window === 'object' && typeof window.document === 'object'
+        ? btoa
+        : (unencoded: string): string => Buffer.from(unencoded).toString('base64')
+
+/**
+ * Extracts custom parameters from a set of SCAPI parameters
+ *
+ * Custom parameters are identified by the 'c_' prefix before their key
+ *
+ * @param parameters object containing all parameters for a SCAPI / SLAS call
+ * @returns new object containing only custom parameters
+ */
+export const extractCustomParameters = (
+    parameters: {[key: string]: string | number | boolean | string[] | number[]} | null
+): helpers.CustomQueryParameters | helpers.CustomRequestBody => {
+    if (typeof parameters !== 'object' || parameters === null) {
+        throw new Error('Invalid input. Expecting an object as an input.')
+    }
+    return Object.fromEntries(Object.entries(parameters).filter(([key]) => key.startsWith('c_')))
 }